pyproject.toml
[build-system]
requires = ["setuptools>=65.5.1", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "llm_quest_benchmark"
version = "0.1.0"
description = "LLM Quest Benchmarking System"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "pandas>=2.0.0",
    "SQLAlchemy>=2.0.0",
    "python-dotenv>=0.19.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "Jinja2>=3.1.2",
    "python-dateutil>=2.8.2",
    "flask>=3.0.0",
    "flask-sqlalchemy>=3.1.0",
    "bootstrap-flask>=2.3.0",
    "flask-wtf>=1.2.0",
    "flask-migrate>=4.0.0",
    "gunicorn>=21.0.0",
    "json-repair>=0.8.0",
    "anthropic>=0.18.0",
    "openai>=1.12.0",
    "openrouter>=0.5.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-timeout>=2.3.0",
    "mypy>=1.0",
    "types-python-dateutil",
    "types-PyYAML",
    "pytest-flask>=1.3.0",
]

[project.scripts]
llm-quest = "llm_quest_benchmark.executors.cli.commands:app"


[tool.setuptools]
packages = ["llm_quest_benchmark"]

[tool.setuptools.package-data]
"llm_quest_benchmark" = [
    "prompt_templates/*.jinja",
    "scripts/*.ts",
    "web/templates/**/*.html",
    "web/static/**/*"
]
